+++ README.md
@@ -1,1 +1,4 @@
 # Project_DA_Java_EN_Come_to_the_Rescue_of_a_Java_Application
+Sample project for OpenClassroom's fundamentals of Java
+
+The program read a list of symptoms from a file and count these symptoms, then print the result of the count by sorting in alphabetical order into another file
